+++ src/services/resources/traits/CreateObjectTrait.php
@@ -31,6 +31,11 @@ trait CreateObjectTrait
     protected abstract static function createRelayBuilderClass(): string;
 
     /**
+     * @return array|TransformerCollectionInterface
+     */
+    public abstract static function defaultTransformer();
+
+    /**
      * @param ObjectBuilderInterface $builder
      * @param ConnectionInterface|string|null $connection
      * @param TransformerCollectionInterface|array|null $transformer
@@ -105,7 +110,7 @@ trait CreateObjectTrait
         TransformerCollectionInterface $transformer = null
     ): PipelineBuilderInterface {
         $transformer = TransformerHelper::populateTransformerCollection(
-            TransformerHelper::resolveCollection($transformer),
+            TransformerHelper::resolveCollection($transformer, static::defaultTransformer()),
             [
                 'resource' => [static::createRelayBuilderClass()]
             ]