+++ src/services/resources/traits/DeleteObjectTrait.php
@@ -33,6 +33,11 @@ trait DeleteObjectTrait
     protected abstract static function deleteRelayBuilderClass(): string;
 
     /**
+     * @return array|TransformerCollectionInterface
+     */
+    public abstract static function defaultTransformer();
+
+    /**
      * @param ObjectBuilderInterface $builder
      * @param ConnectionInterface|string|null $connection
      * @param CacheInterface|string|null $cache
@@ -118,7 +123,7 @@ trait DeleteObjectTrait
         TransformerCollectionInterface $transformer = null
     ): PipelineBuilderInterface {
         $transformer = TransformerHelper::populateTransformerCollection(
-            TransformerHelper::resolveCollection($transformer),
+            TransformerHelper::resolveCollection($transformer, static::defaultTransformer()),
             [
                 'resource' => [static::deleteRelayBuilderClass()]
             ]