+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Removed
+- Deprecated criteria interface methods
+
 ## 1.0.0-rc.2 - 2018-07-03
 ### Changed
 - Altering various Object Association class method names